page_analyzer/pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.0",
    "gunicorn>=21.2.0",
    "psycopg2-binary>=2.9.9",
    "validators>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "flake8>=7.0.0",
    "pytest>=8.0.0",
]
